[GB.QT] * BUG: Fix GotFocus and LostFocus event generation again.
[GB.GTK]
* BUG: Handle focus events the same way as in gb.qt.


git-svn-id: svn://localhost/gambas/branches/2.0@1997 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.gtk/src/gcontrol.h
@@ -128,10 +128,10 @@ public:
 
 // "Signals"
 	void (*onFinish)(gControl *sender); // Special
-	void (*onFocusEvent)(gControl *sender,long type);
-	bool (*onKeyEvent)(gControl *sender,long type);
-	bool (*onMouseEvent)(gControl *sender,long type);
-	void (*onEnterLeave)(gControl *sender,long type);
+	void (*onFocusEvent)(gControl *sender, int type);
+	bool (*onKeyEvent)(gControl *sender, int type);
+	bool (*onMouseEvent)(gControl *sender, int type);
+	void (*onEnterLeave)(gControl *sender, int type);
 	bool (*onDrag)(gControl *sender);
 	bool (*onDragMove)(gControl *sender);
 	void (*onDrop)(gControl *sender);
